refactor: Optimsed codes.
Move private and protected methods or fields to the last of java files.
