+++ simple-jwt-spring-boot-starter/src/main/java/cn/org/codecrafters/simplejwt/autoconfiguration/JjwtTokenResolverAutoConfiguration.java
@@ -66,17 +66,6 @@ import org.springframework.context.annotation.Bean;
 public class JjwtTokenResolverAutoConfiguration {
 
     /**
-     * The GuidCreator instance to be used for generating JWT IDs (JTI).
-     */
-    private final GuidCreator<?> jtiCreator;
-
-    /**
-     * The {@code SimpleJwtProperties} instance containing the configuration
-     * properties for Simple JWT.
-     */
-    private final SimpleJwtProperties simpleJwtProperties;
-
-    /**
      * Constructs a new {@code SimpleJwtAutoConfiguration} instance with the
      * provided SimpleJwtProperties.
      *
@@ -106,5 +95,16 @@ public class JjwtTokenResolverAutoConfiguration {
                 simpleJwtProperties.secret()
         );
     }
+
+    /**
+     * The GuidCreator instance to be used for generating JWT IDs (JTI).
+     */
+    private final GuidCreator<?> jtiCreator;
+
+    /**
+     * The {@code SimpleJwtProperties} instance containing the configuration
+     * properties for Simple JWT.
+     */
+    private final SimpleJwtProperties simpleJwtProperties;
 
 }
